Dodanie groupboxa

+++ WeatherApp/WeatherApp/WeatherInfo.cs
@@ -25,8 +25,8 @@ namespace WeatherApp
         public class Main
         {
             public double temp { get; set; }
-            public int pressure { get; set; }
-            public int humidity { get; set; }
+            public double pressure { get; set; }
+            public double humidity { get; set; }
             public double temp_min { get; set; }
             public double temp_max { get; set; }
         }
@@ -34,7 +34,7 @@ namespace WeatherApp
         public class Wind
         {
             public double speed { get; set; }
-            public int deg { get; set; }
+            public double deg { get; set; }
         }
 
         public class Clouds
